+++ src/core/prelude/regexp/CHANGELOG.md
@@ -9,6 +9,13 @@ Changelog
 > - :house:      [Internal]
 > - :nail_care:  [Polish]
 
+## v3.31.0 (2021-02-03)
+
+#### :rocket: New Feature
+
+* Added new methods `addFlags`, `removeFlags`, `setFlags`
+* Added new static methods `addFlags`, `removeFlags`, `setFlags`, `test`
+
 ## v3.20.0 (2020-07-05)
 
 #### :rocket: New Feature